requirements.txt
streamlit==1.39.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
langchain-core==0.1.40
langchain-openai==0.0.8
langchain-community==0.0.27
docx2txt==0.8
pypdf==3.17.4
langchain-chroma==0.1.4
sentence-transformers==2.2.2
chromadb
openai
langchain==0.1.11
python-multipart
pymupdf